feat: PostSpecification, PostRepository, PostService  - PostSpecification, PostRepository, PostService : Post 다중 조건 검색 기능 추가  - PostService : 단일, 전체 Post 조회 기능 추가 fix: 전체 기능 프로젝트 흐름에 맞춰 수정

+++ src/main/java/com/example/final_project_17team/review/dto/CreateReviewDto.java
@@ -15,6 +15,6 @@ import java.util.List;
 public class CreateReviewDto {
     @NotBlank(message = "리뷰 내용을 작성해주세요.")
     private String content;
-    private Float ratings;
+    private Integer ratings;
     private List<MultipartFile> imageList;
 }
